Java 8 compilation. #PL-5830

+++ modules/gui/test/com/haulmont/cuba/gui/components/TextFieldTest.java
@@ -57,7 +57,7 @@ public abstract class TextFieldTest extends AbstractComponentTest {
         component.setDatatype(Datatypes.getNN(Integer.class));
         component.setValue(10);
 
-        assertEquals(10, component.getValue());
+        assertEquals(10, (int) component.getValue());
     }
 
     @Test
